+++ plugins/gitlab/manifest.go
@@ -7,7 +7,7 @@ import (
 
 const (
 	PluginName        = "gitlab"
-	PluginVersion     = "0.3.1"
+	PluginVersion     = "0.4.0"
 	PluginDescription = "GitLab operations, datasources, indexes, and reverse lookups."
 
 	AuthMethodPersonalAccessToken = "personal_access_token"